# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateRoomRequest:
    properties:
      description:
        maxLength: 500
        type: string
      is_encrypted:
        type: boolean
      name:
        maxLength: 100
        type: string
      topic:
        maxLength: 500
        type: string
      type:
        type: string
    required:
    - type
    type: object
  handler.InviteMemberRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  handler.markReadRequest:
    properties:
      last_read_message_id:
        type: string
    required:
    - last_read_message_id
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Saybridge Platform Backend API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Saybridge API
  version: "1.0"
paths:
...
  /api/v1/rooms:
    get:
      description: Get all rooms the authenticated user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: List of rooms
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Fetch failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Create a new chat room (channel, group, or direct message)
      parameters:
      - description: Room creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Room created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input or creation failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - Rooms
  /api/v1/rooms/{id}:
    get:
      description: Get detailed information about a specific room including members
      parameters:
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room details
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid room ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get room details
      tags:
      - Rooms
  /api/v1/rooms/{id}/leave:
    post:
      description: Leave a room as the authenticated user
      parameters:
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Left room
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Leave failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Leave a room
      tags:
      - Rooms
  /api/v1/rooms/{id}/members:
    post:
      consumes:
      - application/json
      description: Add a user as a member of a room
      parameters:
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User to invite
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Member added
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Invite failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite member to room
      tags:
      - Rooms
  /api/v1/rooms/{id}/members/{userId}:
    delete:
      description: Remove a user from room membership
      parameters:
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User ID to kick (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Member kicked
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Kick failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kick member from room
      tags:
      - Rooms
  /api/v1/rooms/{id}/read:
    post:
      consumes:
      - application/json
      description: Mark all messages in a room as read up to a specific message ID
      parameters:
      - description: Room ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Last read message ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.markReadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Marked as read
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark room as read
      tags:
      - Rooms
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"